+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.9.6] - 2023-11-05
+
+## Changed
+
+- prism-code-editor update
+- scrap-engine update
+
 ## [2.9.5] - 2023-11-03
 
 ### Changed